--- module-common/src/main/java/com/daniel/domain/DTO/order/PaymentActionDetailsDTO.java
@@ -1,12 +1,0 @@
-package com.daniel.domain.DTO.order;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PaymentActionDetailsDTO {
-
-    private String aid, approvedAt, paymentActionType, payload;
-    private Integer amount, pointAmount, discountAmount;
-}
