+++ module-common/src/main/java/com/daniel/domain/DTO/order/request/PayCancelDTO.java
@@ -1,0 +1,19 @@
+package com.daniel.domain.DTO.order.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PayCancelDTO {
+
+    String tid;
+
+    Integer cancelAmount;
+
+    Integer cancelTaxFreeAmount;
+}
