+++ module-common/src/main/java/com/daniel/domain/DTO/order/response/AmountDTO.java
@@ -1,14 +1,18 @@
-package com.daniel.domain.DTO.order;
+package com.daniel.domain.DTO.order.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
+@NoArgsConstructor
 @Getter
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class CancelAvailableAmountDTO {
+@Builder
+public class AmountDTO {
 
-    Integer total, taxFree, vat, point, discount;
+    private Integer total, taxFree, vat, point, discount;
 }
