+++ module-common/src/main/java/com/daniel/domain/DTO/order/response/PayReadyDTO.java
@@ -1,18 +1,21 @@
-package com.daniel.domain.DTO.order;
+package com.daniel.domain.DTO.order.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.Date;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-@Builder
-public class AmountDTO {
+public class PayReadyDTO {
 
-    private Integer total, taxFree, vat, point, discount;
+    private String partnerOrderId, partnerUserId, itemName;
+    private Integer quantity, totalAmount, taxFreeAmount;
+    private String tid, nextRedirectPcUrl;
+    private Date createdAt;
 }
