+++ module-common/src/main/java/com/daniel/domain/DTO/order/response/kakaoPay/ApprovedCancelAmount.java
@@ -1,0 +1,18 @@
+package com.daniel.domain.DTO.order.response.kakaoPay;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+@Builder
+public class ApprovedCancelAmount {
+
+    private Integer total, taxFree, vat, point, discount;
+}
